DotNet: Allow Dynamic/Provider wide events on Linux
Dynamic/Self describing events do not use a static ID, they use a name
along with a format as the true "identity" of the event. We also need to
be able to do things on a provider wide basis, such as recording all
events from a provider for keyword(s) and level(s).

Add default_tracepoint support to Linux IPC to ensure non-described
event IDs will go to a specific tracepoint.

Add name lookup for self describing events on the default_tracepoint
handler. This ensures callers do not have to put an ID, for now ID 0 is
used for this. When ID is 0 we treat it as dynamic/self describing. In
future commits this will be more specific at the scripting level.

Add a list of provider wide callbacks on Linux IPC provider. These will
always run for all registered tracepoints. This ensures we can have a
single list of callbacks that ALWAYS get run for a given provider.
Future commits will leverage this to record keyword and level events
across a provider without requiring unique IDs.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>