DotNet: Add record_dotnet_provider() script method
Sometimes people want to record a provider within DotNet at a keyword +
level basis, instead of a per-event basis.

Add record_dotnet_provider() script method to allow for this.

Add new_dotnet_provider_flags() script method to create a flag struct
that can be passed to record_dotnet_provider(). The flags allow
specifying if callstacks should be collected and for which keywords.

Add ways for trace events to dynamically register kind and record_types
as well as disable callstacks (only first IP). This is required when we
have to define a single tracepoint to get callstacks, but want to
specify only some events within that tracepoint should take the space to
save the full callstack.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>