record-trace: Move core logic into engine module
We want to allow record-trace to be called directly via an FFI in the
future. In order to do this, we must have hte core logic be available
outside of the CLI.

Move core logic into an engine crate.

Add EngineOutput struct to allow handling when and how text should be
displayed in different contexts (error, normal, live, etc).

Use EngineOutput in CLI to stop when control-c is hit and output
everything else to default callbacks (stdout and stderr).

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>