record-trace: Add FFI library
Tools that would like to incorporate record-trace/one-collect but
are not written in Rust need a shared object (.so/.dll).

Add FFI sub-directory and define "RecordTrace" C API for non-Rust
languages to utilize.

The API is very simple and expandable with minimal ties between Rust and
non-Rust languages. The same command line passed to the record-trace
tool is passed to the RecordTrace API along with a callback. The
callback is given the type of message and the message bytes. If non-zero
is returned from the callback, the recording stops.

Add record-trace/ffi to github CI build/test tasks.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>