Event: Add LocationType::StaticLenPrefixArray
Some events, such as those from DotNet, utilize arrays of data that are
variable length at static locations. The length of the array is length
encoded as a 16-bit value before the elements. The type of the elements
are known via the field type and are not encoded in the payload.

Add LocationType::StaticLenPrefixArray to allow these events to be
decoded properly by other file formats, closures, etc.

Handle LocationType::StaticLenPrefixArray in both skipping and data
fetching closure logic within Event struct.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>