NetTrace: Add Array Support
We now support 16-bit length prefixed arrays, which NetTrace also
supports. We need to enable detecting these from EventFields and then
writing out the correct metadata to ensure proper decoding.

Add element_type_id to NetTraceField and update existing fields to set
it to None.

Add logic to find element type for an EventField struct, utilize it for
non-Array fields as well as for array fields.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>