Exporting: Add span/trace_id record attributes
The W3C TraceContext is very important for HTTP workloads, we need to
support this ideally at an OS level, but also at a per-event basis. We
only currently support strings and u64 attributes, for TraceContext and
other eventual attributes, we need to be able to store byte arrays.

Add new ExportAttributeValue::Record, which stores a record ID that contains
bytes with an optional format of those bytes for later decoding.

Store TraceId and SpanId attributes at the OS level if they are defined
from the underlying sessions (or overrides) as Record attributes.

Add methods to create records from ExportTraceContext structs during
callbacks, etc. This includes overrides, like we do for other OS
attributes. This allows per-event capturing of these details that the OS
might not know about.

Move default attribute creation from a single function into a list of
ExportAttributeSource trait structs. Expose these sources as options in
ExportSettings. This allows tools and formats to define what attributes
are collected by default for each event. This is useful for formats that
do not understand these attributes (or do not have any attribute
concepts). For example, stdout formats that never save the actual data.
It's not useful to save all this extra data for it to get thrown out at
the end.

Limit attribute creation exposure by making new_*() within
ExportAttributePair to only crate public. This is required because
records need more coordination than should be exposed to non-crate code.

Expose creating records via ExportMachine to allow non-crate code to
create Record attributes if needed instead of at the ExportAttributePair
level. This is a much safer design to expose to non-crate code, as we
can evolve it vs ExportAttributePair::new_*() methods (that are now only
crate visible).

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>