record-trace: Add full set of attributes
The ExportSettings struct now allows configuration of which attributes
to save for all events. Since we support a wide array of formats, we
need to enable most of these by default.

For now, enable all attributes by default. Eventually the format we pick
should decide which of these attributes are needed.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>