DotNet: Ensure right format ID for dynamic events
We need to differentiate between events that are named and not named
when caching the format ID on Linux. When non-dynamic events come in,
the metadata may be present, but if there is no name associated with
it, we must treat it as a manifest event.

Update format_key to use low 31 bits for the ID, and the top bit to
indicate named or not named when doing lookups/caching of format details
(including the event ID).

Do not set name_id to Some when the event name is empty.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>