Events: Add extra bounds checking
Rust will panic if a Range of a slice starts outside of the slice. While
this prevents out of bounds from being a security issue, it does cause
the entire process to exit. This is not ideal.

Add bounds checking to Event data fetching methods and give back an
error instead of relying upon Panics. This ensures one bad apple does
not spoil the bunch when collecting data.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>